+++ Coin/Model/QuickStatistic.swift
@@ -9,11 +9,11 @@ import Foundation
 
 class QuickStatistic: Decodable {
     var currency: String
-    var totalRemainder: Double
-    var totalExpense: Double
-    var totalBudget: Double
+    var totalRemainder: Decimal
+    var totalExpense: Decimal
+    var totalBudget: Decimal
     
-    init(currency: String = "USD", totalRemainder: Double = 0, totalExpense: Double = 0, totalBudget: Double = 0) {
+    init(currency: String = "USD", totalRemainder: Decimal = 0, totalExpense: Decimal = 0, totalBudget: Decimal = 0) {
         self.currency = currency
         self.totalRemainder = totalRemainder
         self.totalExpense = totalExpense
